Convert user-provided values to floats
These values don't need double precision, only the internal calculated
values do (might?).

+++ include/cpd/affinity_eigenvectors.hpp
@@ -30,7 +30,7 @@ namespace cpd
 {
 
 
-void find_affinity_eigenvectors(const arma::mat& Y, double beta,
+void find_affinity_eigenvectors(const arma::mat& Y, float beta,
                                 const arma::uword numeig, arma::mat& Q, arma::mat& S);
