+++ include/cpd/find_P.hpp
@@ -34,7 +34,7 @@ double find_P(
         const arma::mat& X,
         const arma::mat& Y,
         double sigma2,
-        double outliers,
+        float outliers,
         arma::vec& P1,
         arma::vec& Pt1,
         arma::mat& PX,
